requirements.txt
pandas
numpy
plotly
dash
dash-bootstrap-components
Pillow==9.3.0
python-dotenv
gunicorn
psycopg2-binary
